+++ kernel/kmain.h
@@ -20,6 +20,6 @@
 #include "multiboot.h"
 #include <kernel/kstd/unix_types.h>
 
-struct multiboot_info parse_mboot(uint32_t physaddr);
-extern "C" int kmain(uint32_t mbootptr);
+struct multiboot_info parse_mboot(size_t physaddr);
+extern "C" int kmain(size_t mbootptr);
 void kmain_late();
